Add inactive_focused_workspace support
This new i3bar configuration allows a user to configure the color,
border and background of a workspace button that is currently inactive
_and_ focused - "focused" meaning that the workspace is currently under
the focused monitor.

Unlike previously, this makes its configuration distinct from the
inactive_workspace button. It is especially useful when the bar color is
different on a focused and unfocused monitor.

By default the inactive_focused_workspace colors remain the same as the
inactive_workspace colors, so the change won't be notived unless
explicitely configured.

Signed-off-by: Antonin Godard <antoningodard@pm.me>